refactor file and function names
